refactor: correct according to the check
