+++ src/shared/utils/index.ts
@@ -1,2 +1,2 @@
 export { getDefaultTheme } from './get-default-theme';
-export { localStorageController } from './localstorage-controller';
+export { localStorageController } from './local-storage-controller';
